+++ packages/react/src/webgpu/index.ts
@@ -1,0 +1,2 @@
+export * from './dotlottie';
+export type * from '@lottiefiles/dotlottie-web/webgpu';