+++ packages/web/src/software/index.ts
@@ -1,0 +1,3 @@
+export * from '../common';
+export * from './dotlottie';
+export * from './worker';